+++ src/sdk/PnP.Core.Test/Base/DeleteTests.cs
@@ -50,7 +50,9 @@ namespace PnP.Core.Test.Base
                 var listCount = context.Web.Lists.Length;
 
                 // Delete the list
-                await myList.DeleteAsync();
+                await myList.WithSPResponseHeaders((responseHeaders) => { 
+                    Assert.IsTrue(!string.IsNullOrEmpty(responseHeaders["SPRequestGuid"])); 
+                }).DeleteAsync();
                 // Verify that the list was removed from the model collection as well
                 Assert.IsTrue(context.Web.Lists.Length == listCount - 1);
 
@@ -100,11 +102,17 @@ namespace PnP.Core.Test.Base
                 // Delete the list
                 var batch = context.NewBatch();
 
-                await myList.DeleteBatchAsync(batch);
+                // Shows how to get the response headers from the batch for the SharePoint REST and CSOM requests
+                await myList.WithSPResponseHeaders((responseHeaders) => {
+                    Assert.IsTrue(!string.IsNullOrEmpty(responseHeaders["SPRequestGuid"]));
+                }).DeleteBatchAsync(batch);
                 await context.ExecuteAsync(batch);
 
                 // Check that the batch response was OK
                 Assert.IsTrue(batch.Requests[0].ResponseHttpStatusCode == HttpStatusCode.OK);
+
+                // Shortcut to just get the SPRequestGuid from the batch response
+                Assert.IsTrue(batch.Requests[0].SPRequestGuid != null);
 
                 // Was the list added
                 bool exceptionThrown = false;